pyproject.toml
[tool.poetry]
name = "es-adry-pia"
version = "1.0.0"
description = "Artificial Intelligence Programming Subject"
authors = ["Adrián Mateos Pérez"]
license = "GPL-3.0 license"
readme = "README.md"
packages = [{include = "tasks"}, {include = "exercises"}]

[tool.poetry.scripts]
task1 = "tasks.task1:run"
rps = "exercises.RPS_Basic:main"
stateMachine = "exercises.StateMachine:main"
practica01 = "practicaguiada01.api:main"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
boto3 = "^1.28.77"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
opencv-python = "^4.8.1.78"
numpy = "^1.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
